requirements.txt
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.5.0
PyYAML==6.0
quaternion==3.5.2.post4
scikit_learn==1.2.2
scipy==1.10.1
timm==0.6.13
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
tensorboard==2.12.2